Step R73: five of R71's six adoptable bases taken (#56)
`failed` 117 -> 125, `total` 29032 -> 29077, compliance missing 18 -> 13. A deliberate rise on
the owner's instruction: +45 tests, +37 green, +8 red, and all eight reds were named before the
run. FIXED none, BROKEN exactly those eight.

Three of the five are tier moves rather than additions, since a base belongs to one tier. The
re-parent lets NonSharedPrimitiveCollections' hand-mirrored Array_of_byte override go.

The trap on the AdHocMiscellaneous move: four of the six reds R71 measured are on the CORE base
and pass at Tier A. The Tier A class carried EF's InMemory overrides for them, and those had to
move too -- they assert the size of EF's relational command cache, and it is the CLIENT's cache
they read. This client is not relational on either tier, so the backing store does not change the
reason. Carrying them turns R71's "+7 green / 6 red" into the true +11 / 2.

R62's SetParameterizedCollectionMode blocker does not exist: a no-op is enough, because only
tests asking for a non-default mode consult it. ToSqlQuery was priced as blocked twice over and
is blocked by neither.

Also corrects R71. Its claim that UdfDbFunction's 80 reds are SQLite's missing functions was
wrong, and reading the reasons rather than the base's name is what showed it: 2 of 80 are the
store, 56 are this provider refusing or evaluating a HasDbFunction call, and 11 run and get it
wrong. R62 was right about that mechanism. Those 11 are diagnosed before the base is adopted.
