+++ test/known-failures.names.txt
@@ -18,6 +18,8 @@ InfoCarrier.Core.FunctionalTests.Sqlite.DataAnnotationInfoCarrierTest.Table_can_
 InfoCarrier.Core.FunctionalTests.Sqlite.EntitySplittingInfoCarrierTest.ExecuteDelete_throws_for_entity_splitting(async: False)
 InfoCarrier.Core.FunctionalTests.Sqlite.EntitySplittingInfoCarrierTest.ExecuteDelete_throws_for_entity_splitting(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.AdHocAdvancedMappingsQuerySqliteInfoCarrierTest.Casts_are_removed_from_expression_tree_when_redundant
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.AdHocMiscellaneousQuerySqliteInfoCarrierTest.Multiple_different_entity_type_from_different_namespaces(async: False)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.AdHocMiscellaneousQuerySqliteInfoCarrierTest.Multiple_different_entity_type_from_different_namespaces(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.Associations.OwnedJsonStructuralEqualityQueryInfoCarrierTest.Associate_with_parameter_null
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.Associations.OwnedJsonStructuralEqualityQueryInfoCarrierTest.Contains_with_nested_and_composed_operators
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.Associations.OwnedJsonStructuralEqualityQueryInfoCarrierTest.Contains_with_operators_composed_on_the_collection
@@ -36,6 +38,12 @@ InfoCarrier.Core.FunctionalTests.Sqlite.Query.JsonQuerySqliteInfoCarrierTest.Fro
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.JsonQuerySqliteInfoCarrierTest.FromSql_on_entity_with_json_project_json_collection(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.JsonQuerySqliteInfoCarrierTest.FromSql_on_entity_with_json_project_json_reference(async: False)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.JsonQuerySqliteInfoCarrierTest.FromSql_on_entity_with_json_project_json_reference(async: True)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.NonSharedPrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_Contains_with_default_mode_EF_MultipleParameters(mode: Constant)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.NonSharedPrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_Contains_with_default_mode_EF_MultipleParameters(mode: MultipleParameters)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.NonSharedPrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_Contains_with_default_mode_EF_MultipleParameters(mode: Parameter)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.NonSharedPrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_Count_with_column_predicate_with_default_mode_EF_MultipleParameters(mode: Constant)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.NonSharedPrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_Count_with_column_predicate_with_default_mode_EF_MultipleParameters(mode: MultipleParameters)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.NonSharedPrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_Count_with_column_predicate_with_default_mode_EF_MultipleParameters(mode: Parameter)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.NorthwindSetOperationsQueryInfoCarrierTest.Collection_projection_before_set_operation_fails(async: False)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.NorthwindSetOperationsQueryInfoCarrierTest.Collection_projection_before_set_operation_fails(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.NullSemanticsQueryInfoCarrierTest.CaseOpWhen_predicate(async: False)
